# docs/swagger.yaml
definitions:
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  domain.Role:
    enum:
    - participant
    - staff
    - admin
    type: string
    x-enum-varnames:
    - RoleParticipant
    - RoleStaff
    - RoleAdmin
  domain.TokenResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      userId:
        type: string
    type: object
  domain.User:
    properties:
      email:
        type: string
      foodLimitation:
        type: string
      id:
        type: string
      imageURL:
        type: string
      invitationCode:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/domain.Role'
      sizeJersey:
        type: string
      state:
        type: string
      university:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/users:
    get:
      description: Retrieve a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.User'
            type: array
        "500":
          description: Failed to fetch users
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get all users
  /api/users/{id}:
    get:
      description: Retrieve a user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Failed to fetch user
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get user by ID
    patch:
      consumes:
      - application/json
      description: Update a user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Failed to update user
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user by ID
  /api/users/register:
    post:
      consumes:
      - application/json
      description: Register a new user in the system
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.TokenResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Failed to create user
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register a new user
swagger: "2.0"